Add SequenceHandler draft
The SequenceHandler is desigent to handel complex chains of actions with progress to achive a function asynchronosly. Each step of the process can be order by a sequence number, given a relavite weight for progress, be vital or caneled on its own.

Events for start and end of each step aswell as the whole sequence are triggered for interaction with other code, mostly UI.