Remove static components form Profession
With this commit the remaining static components for stroing, providing and creating Profession
instances are now removed and replaces with the instanced version of ProfessionRepository
and ProfessionDataBuilder.

This completes the transition in for the static GW2 data from in class stattic storage to "external"
storage classes with indpendend instanced objects.