Add interface abstraction layer to GameAreas and Enemies
This is necessary since the new Log interfeaces need to reference boss -> this is better done in the interfaces
project than the data project, since the data project will no longer create / hold any concrete
implemenation of the Log data tpyes.

The Logtypes will be implemented internally by the database for storate  and prbably the
LogUploader.Tools or LogUploader.EliteInsights project for new data processing.