Add support for the RO+ lieutenants feature
Lieutenants are a new feature of RO+, Lieutenants can do all relevent stuff to update a raid.
Every raid where the user is a lieutenants will now be shwen as a possible raid to update.

Implementation:
The role number for lieutenants is 1, so the check was ajusted from role == 2 to role >= 1.