Create concept of EliteInsights Json processors
IEiProcessor is a class wich taks the json output of Elite Insights fom a speciefied range of EiVersion and
processes it into a Full log for further usage.

This approche enable the developers the be backward and mostly forward compatible with the Elite
Insights json parsing wile keeping the maintenence mostly side effect free since every version of EI
could have a totally unique and indipenden parsing.

In this first commit only the shells of the minimal backward supported EI verions are created.
Implementation of the Parserrs will follow up in the next commits.

Note:
The tests are in the LongRunning category and marked Explicite to keep test time in ckeck.