Fix bug with default discord post avatar url not set correctly
Due to refactoring the discrod prost avatar url was set to the UI placeholder string "<Default>" instad of the default
url to the image on the web.

Fixed by adding a property for direct access the privat field, which stores the correct value all the time. Also updated the
read call of the AvatarUrl to the new property.