Change RO+ role update to not loose additonal roles
The multi role support for RO+ in the initial verison only did an override of the first role, however this could lead to
situations of doubled roles, a Tank/Heal was changed to Heal/Heal. This is not help full, to  cope with this,
whil keeping the first element the one the locic of role updates works on this change was made.

The new behaviour says if the role that is tried to set by the guess role logic is already present in the roles of the
positon, swap the places/order of the first and the found one.

This ensures roles that already existed, will not be removed on performing a RO+ update.