Improve memory usage for EI json parsing
The perofrmace of the parsing of the json output of EI was
improved by using a custom JsonReader, that is capeable of
sipping userdefinded tokens and preventing these to get to
the inefficent JObject.Load implementation.

Depending on the data this reduces the memeory usage
spike by half or more.