Fix crash when parsing of simpleLog.json faild
The parsing of the simple log json could break if a charakter name
containd a charakter that was not part of the iso-latin-1 charset. The
resulting enconding missmatch lead to an invalid json file. The
exception from this was not handled an caused the programm to crash
hard.

This is fixed by surrounding the code responsible for parsing in a
try-catch block to catch the exception, log it, delete the invalid
json and remove it from the database.

The only downside to this fix is any files that are corrupted are
deleted and therefor the preview of the affected log is borken.

The underlaying encoding issue will be solved with the next commit.