Change json file encoding to UTF-8 with BOM
Json files were encoded using iso-latin-1 encoding. This however makes
issues, since GW2 allows charakters with names that contain characters
not containd in this charset.

To fix this UTF-8 with BOM is used from now on for the chaching json
files. UTF-8 supports the necessary characters for the case described
above.

To keep this change backward compatible the BOM is prepended to all new
UTF-8 encoded files. This allows to tell apart if a file was written
using the old or new way, so that these can properly be converted into
a string.