Rework RO+ role detection and player matching
The role detection and player matching for RO+ was lacking behind. Missing features were the new Roles, Alacrity and Quickness, as well as the mulitple roles per position. Further there was a Bug with LFG players not being properly updated when already marked as LFG by the user in RO+.

To address these issues the logic behind assigening roles and matching players form log to RO+ was reworked. Because the old implemenation heavily relied on the fact that every player could only have one role, the only proper solution was to rework the entiere thing.

Role detection:
The new role detection is divided into 3 parts. The first part handles each class on it's own, handling dps type, boon support and heal. In the second step all Players and the boss are considerd, this assigens the Tank, Banners, Kiters and special roles for each boss. In the end all players are validated to only have a max of 4 roles.

Player assigment:
The basic idea stayed the same, however the operation was made more "destruktive" and will more often tust the Guessed roles than the roles set by the user in RO+. Further the matchin is now aware of the multiple roles and the assigment bug should be fixed.