Add support for rate limiting of dps.report
dps.report only allows for 10 uploads per 60 seconds. To slow down the
uploade with minimal impact for the user. Before each upload a 7s
background delay is started, the next upload may only begin after the
delay. This ensure the rate limiting of dps.report is fufilled and does not
cause erros here.