Fix SQL injection for bulk insert.
As the original idea behind bulk insert was to reduce the overhead for running many single inserts by combining them in one large sql command. As this used string concat this was not injection safe.

To achieve a middle ground, now the one insertion per command is performed, however all insertions are wrapped inside of a single transaction to reduce overhead of a transaction per query.