Fix boss detection based on folder name
The arcdps updates from jun.04.2024 and jun.08.2024 changed the default names of the log folder to: "map name - npc name" and "npc name (id)" respectively.

This broke the detection as full string equality was required for detection.

To fix this, the folder name now only has to contain one of the folder names specified in the `DataConfig.json`.
This allows for both of this changes to be properly detected while being compatible with the old behavior.