Fix new logs being added multiple times to the database
The change to use proper dapper parameters instead of format string caused this. This was because Dapper parameters of strings Automatically add quotes to strings and do not need them provided manually. As these were still present. this caused the logs to not be found by path. This lead the check if the log was recored to fail and re added the log therefor.